// reference-code/esr/esale-mobile/src/shared/components/suggestions/interface/task-suggest-interface.ts
export interface MilestoneTaskSuggest {
  milestoneId: number;
  milestoneName: string;
}


/**
 * Define values of Customer
 */
export interface CustomerTaskSuggest {
  customerId: number;
  customerName: string;
  parentCustomerName: string;
}

/**
 * Define values of Product
 */
export interface ProductTradings {
  productTradingId: number,
  productTradingName: string
}

/**
 * Define values of TaskSuggest
 */
export interface Tasks {
  taskId: number,// data in database mapping data form
  taskName: string,// data in database mapping
  milestone: MilestoneTaskSuggest,// data in database mapping
  customer: CustomerTaskSuggest,// data in database mapping
  productTradings: ProductTradings[],// data in database mapping
  finishDate: string,// data in database mapping
  status: number// data in database mapping
  operators: {
    employeeId: number,
    employeeName: string,
    employeeSurname: string
  }[]
}

/**
 * Define values of TaskSuggest
 */
export interface TaskSuggest {
  taskId: number,// data in database mapping data form
  parentCustomerName: string, // data in database mapping
  milestoneName: string,// data in database mapping
  customerName: string,// data in database mapping
  productTradingName: string,
  taskName: string,// data in database mapping
  finishDate: string,// data in database mapping
  finishType: number,// data in database mapping
  employeeName: string,// data in database mapping
  status: number,// data in database mapping
  data: any
}